refactor(showInfo): 将buildInfo提取为独立脚本
脚本在 /src/prebuild.js 处，在 vue.config.js 中通过 require 引用并执行
